% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex-hull.R
\name{convex_hull}
\alias{convex_hull}
\title{Convex hull}
\usage{
convex_hull(points)
}
\arguments{
\item{points}{The 10,000 points data set}
}
\description{
Plot the convex hull for the 10,000 points data set
}
\examples{
convex hull()
}
