% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdtree.R
\name{kdtree}
\alias{kdtree}
\title{k-d tree}
\usage{
kdtree(points)
}
\arguments{
\item{points}{A data frame with columns for x and y coordinates, and each point in a row.}
}
\description{
Compute a k-d tree for a given set of points
}
\examples{
kdtree()
}
\keyword{k-d}
\keyword{tree}
