% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points-data.R
\docType{data}
\name{points}
\alias{points}
\title{10,000 points data set}
\format{Data frame}
\usage{
data(points)
}
\description{
10,000 random uniform points on a 10,000 x 10,000 grid
}
\examples{
data(points)
}
\keyword{datasets}
