% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weiszfeld.R
\name{weiszfeld}
\alias{weiszfeld}
\title{Weiszfeld}
\usage{
weiszfeld(points, p0)
}
\arguments{
\item{points}{A data frame with columns for x and y coordinates}

\item{p0}{A vector with the (x, y) co-ordinates of the initial point}
}
\description{
Solves the Fermat problem using Weiszfeld's algorithm
}
\examples{
weiszfeld()
}
\keyword{weiszfeld}
